requirements.txt
ffmpeg-python
numpy
mediapipe
opencv-python
psutil